requirements.txt
agate==1.9.1
aiobotocore==2.19.0
aiofiles==24.1.0
aiohttp==3.9.3
aioitertools==0.12.0
aiosignal==1.3.1
alembic==1.13.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.3.0
apache-airflow==2.9.0
apache-airflow-providers-amazon==9.2.0
apache-airflow-providers-apache-spark==5.0.0
apache-airflow-providers-celery==3.6.1
apache-airflow-providers-cncf-kubernetes==10.1.0
apache-airflow-providers-common-compat==1.3.0
apache-airflow-providers-common-io==1.3.0
apache-airflow-providers-common-sql==1.21.0
apache-airflow-providers-fab==1.0.2
apache-airflow-providers-ftp==3.7.0
apache-airflow-providers-http==4.10.0
apache-airflow-providers-imap==3.5.0
apache-airflow-providers-smtp==1.6.1
apache-airflow-providers-sqlite==3.7.1
apispec==6.6.0
argcomplete==3.2.3
asgiref==3.8.1
asn1crypto==1.5.1
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.13.3
billiard==4.2.0
blinker==1.7.0
boto3==1.36.12
boto3-stubs==1.36.18
botocore==1.36.3
botocore-stubs==1.36.18
cachelib==0.9.0
cachetools==5.5.1
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.3
croniter==2.0.3
cryptography==41.0.7
daff==1.3.46
dbt-adapters==1.13.2
dbt-athena-community==1.8.4
dbt-common==1.14.0
dbt-core==1.9.1
dbt-extractor==0.5.1
dbt-postgres==1.9.0
dbt-semantic-interfaces==0.7.4
deepdiff==7.0.1
Deprecated==1.2.14
dill==0.3.1.1
distlib==0.3.9
dnspython==2.6.1
docutils==0.16
durationpy==0.9
elastic-transport==8.17.0
elasticsearch==8.17.1
email_validator==2.1.1
filelock==3.17.0
Flask==2.2.5
Flask-AppBuilder==4.4.1
Flask-Babel==2.0.0
Flask-Caching==2.1.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.1
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
flower==2.0.1
frozenlist==1.4.1
fsspec==2025.2.0
google-auth==2.38.0
google-re2==1.1
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.70.0
grpcio-status==1.62.3
gunicorn==21.2.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
humanize==4.9.0
idna==3.6
importlib-metadata==6.11.0
importlib_resources==6.4.0
inflection==0.5.1
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
jsonpath-ng==1.7.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.6
kubernetes==31.0.0
kubernetes_asyncio==30.3.1
lazy-object-proxy==1.10.0
leather==0.4.0
limits==3.10.1
linkify-it-py==2.0.3
lockfile==0.12.2
lxml==5.3.0
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mashumaro==3.14
mdit-py-plugins==0.4.0
mdurl==0.1.2
mmh3==4.1.0
more-itertools==10.2.0
msgpack==1.1.0
multidict==6.0.5
mypy-boto3-athena==1.36.0
mypy-boto3-glue==1.36.4
mypy-boto3-lakeformation==1.36.0
mypy-boto3-sts==1.36.0
networkx==3.4.2
numpy==2.2.2
oauthlib==3.2.2
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
ordered-set==4.1.0
packaging==24.0
pandas==2.2.3
parsedatetime==2.6
pathspec==0.12.1
pendulum==3.0.0
platformdirs==4.3.6
pluggy==1.4.0
ply==3.11
prison==0.2.1
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==5.29.3
psutil==5.9.8
psycopg2-binary==2.9.10
py4j==0.10.9.7
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAthena==3.12.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.17.2
PyJWT==2.8.0
pyspark==3.5.4
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-nvd3==0.15.0
python-slugify==4.0.1
python3-saml==1.16.0
pytimeparse==1.1.8
pytz==2024.1
PyYAML==6.0.1
redshift-connector==2.1.5
referencing==0.31.1
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.7.1
rich-argparse==1.4.0
rpds-py==0.18.0
rsa==4.9
s3fs==2025.2.0
s3transfer==0.11.2
scramp==1.4.5
setproctitle==1.3.3
setuptools==75.8.0
six==1.16.0
sniffio==1.3.1
snowplow-tracker==1.0.4
soupsieve==2.6
SQLAlchemy==1.4.52
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
text-unidecode==1.3
time-machine==2.14.1
tornado==6.4
types-awscrt==0.23.10
types-requests==2.32.0.20241016
types-s3transfer==0.11.2
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.2
urllib3==2.3.0
vine==5.1.0
virtualenv==20.29.1
watchtower==3.3.1
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==2.2.3
wheel==0.45.1
wrapt==1.16.0
WTForms==3.1.2
xmlsec==1.3.14
yarl==1.9.4
zipp==3.18.1
